requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
idna==3.6
paseto==1.0.1
pendulum==2.1.2
pip==23.0.1
pycryptodomex==3.19.1
pydantic==2.5.3
pydantic-core==2.14.6
pymongo==4.6.1
pysodium==0.7.17
python-dateutil==2.8.2
pytzdata==2020.1
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing-extensions==4.9.0
uvicorn==0.25.0
wheel==0.42.0
python-jose
python-dotenv